requirements.txt
pipecat-ai[assemblyai,cartesia,openai,daily,silero]>=0.0.47
python-dotenv>=1.0
loguru>=0.7
